pyproject.toml
[project]
name = "nslab2d"
version = "0.1.0"
description = "Python port of DNSLab NS2dLab with CPU and optional GPU backends"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy>=2.4.4",
  "scipy>=1.17.1",
  "pyfftw>=0.15.1",
  "matplotlib>=3.10.8",
]

[project.optional-dependencies]
gpu = [
  "cupy-cuda12x>=13.4.0",
  "nvidia-cuda-nvrtc-cu12>=12.9.86",
  "nvidia-cufft-cu12>=11.4.1.4",
  "nvidia-nvjitlink-cu12>=12.9.86",
]
dev = [
  "pytest>=9.0.3",
]

[project.scripts]
nslab2d = "nslab2d.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
